refactor: extract mapLane into shared feature lib
Phase 0.5 prep. The mapLane function (lane Firestore doc → Lane type,
including the critical pre-S29 sceneNumber fallback that derives
sceneNumber = sortOrder + 1 for legacy lanes missing the field) was
inlined at the top of useLanes.ts. Moving it into its own module at
src-vnext/features/shots/lib/mapLane.ts so callers outside the hook
(e.g. the forthcoming Cmd+K lazy-index hook) can import the same
defensive normalization instead of rolling an unsafe cast on raw doc
data. Zero behavioral change — useLanes.ts imports the new export.
